backend-flask/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiohttp-retry==2.9.1
aiosignal==1.4.0
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click>=8.1.3,<8.2
colorama==0.4.6
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-MySQL==1.6.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
greenlet==3.2.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
mysql-connector-python==9.5.0
propcache==0.4.1
PyJWT==2.10.1
PyMySQL==1.1.2
requests==2.32.5
SQLAlchemy==2.0.44
twilio==9.8.4
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.22.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
gtts>=2.4.0
Pillow>=10.0.0
sentence-transformers
scikit-learn
